+++ include/internal/libspdm_crypt_lib.h
@@ -1,0 +1,10 @@
+/**
+ *  Copyright Notice:
+ *  Copyright 2021-2022 DMTF. All rights reserved.
+ *  License: BSD 3-Clause License. For full text see link: https://github.com/DMTF/libspdm/blob/main/LICENSE.md
+ **/
+
+#include "library/spdm_crypt_lib.h"
+#include "hal/library/cryptlib.h"
+#include "hal/library/debuglib.h"
+#include "hal/library/memlib.h"
